+++ GODInventoryWinForm/SelectProductForm.cs
@@ -84,10 +84,10 @@ namespace GODInventoryWinForm
             }
 
         }
-        private int GetGenreId()
+        private short GetGenreId()
         {
 
-            return ((this.listBox1.SelectedIndex >= 0) ? (int)this.listBox1.SelectedValue : 0);
+            return (short)((this.listBox1.SelectedIndex >= 0) ? this.listBox1.SelectedValue : 0);
         }
 
         private void listView1_Click(object sender, EventArgs e)
